# internal/apiv1/openapi.yaml
openapi: 3.1.0
info:
  title: wpcalc API v1
  version: "1.0.0"
  description: |
    A stateless JSON REST mirror of wpcalc's server-rendered app, for
    programmatic clients. Authenticated with a bearer API token (`wpcalc
    token create <username>` on the CLI), not the browser session cookie —
    there is no login/logout/tenant-switch operation here, because a bearer
    token is not a session: every request names its tenant explicitly in
    the path instead of relying on a server-side "active tenant".

    This document is generated code's source of truth (`internal/apiv1` is
    generated from it via oapi-codegen) and is itself served by the API at
    `GET /api/v1/openapi.json` and `GET /api/v1/openapi.yaml`.

    Authorization mirrors the RBAC96 model exactly: every operation below
    carries `x-required-permission`/`x-permission-scope`, checked the same
    way as the HTML app (`Identity.CanSystemWide`/`CanInTenant`/`Can`) —
    see `internal/httpx/openapi.yaml` (served at `GET /openapi.json` /
    `/openapi.yaml` / `/openapi.html` by the HTML app) for the full RBAC96
    write-up.
  license:
    name: MIT
    url: https://github.com/rolsim/wpcalc/blob/main/LICENSE
servers:
  - url: /api/v1
security:
  - bearerToken: []
paths:
...
  /tenants/{tenantId}/employees:
    parameters:
      - $ref: "#/components/parameters/TenantId"
    get:
      operationId: listEmployees
      summary: List a tenant's employees
      x-required-permission: manage_employees
      x-permission-scope: tenant
      responses:
        "200":
          description: Every employee in the tenant.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Employee" }
        default: { $ref: "#/components/responses/Error" }
    post:
      operationId: createEmployee
      summary: Create an employee
      x-required-permission: manage_employees
      x-permission-scope: tenant
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EmployeeCreate" }
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Employee" }
        default: { $ref: "#/components/responses/Error" }
  /tenants/{tenantId}/employees/{employeeId}:
    parameters:
      - $ref: "#/components/parameters/TenantId"
      - $ref: "#/components/parameters/EmployeeId"
    get:
      operationId: getEmployee
      summary: Fetch one employee
      x-required-permission: manage_employees
      x-permission-scope: tenant
      responses:
        "200":
          description: The employee.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Employee" }
        default: { $ref: "#/components/responses/Error" }
    patch:
      operationId: updateEmployee
      summary: Update an employee
      x-required-permission: manage_employees
      x-permission-scope: tenant
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EmployeeUpdate" }
      responses:
        "200":
          description: Updated.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Employee" }
        default: { $ref: "#/components/responses/Error" }
    delete:
      operationId: deleteEmployee
      summary: Delete an employee
      description: Cascades to every recorded hour of theirs. No undo.
      x-required-permission: manage_employees
      x-permission-scope: tenant
      responses:
        "204":
          description: Deleted.
        default: { $ref: "#/components/responses/Error" }
...
  /tenants/{tenantId}/months/{ym}/entries:
    parameters:
      - $ref: "#/components/parameters/TenantId"
      - $ref: "#/components/parameters/YearMonth"
    put:
      operationId: setHours
      summary: Record one cell's hours
      description: "`hours: 0` clears the cell."
      x-required-permission: write
      x-permission-scope: employee
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SetHoursRequest" }
      responses:
        "200":
          description: Stored; recomputed totals returned.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/HoursResult" }
        default: { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    bearerToken:
      type: http
      scheme: bearer
      description: >-
        An API token minted by `wpcalc token create <username>`. Sent as
        `Authorization: Bearer <token>`. Tokens are opaque, hashed at rest
        (SHA-256), shown in full only once (at creation time), and can be
        revoked individually without touching the account's password or any
        browser session.
  parameters:
    TenantId:
      name: tenantId
      in: path
      required: true
      schema: { type: string, pattern: "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$" }
    EmployeeId:
      name: employeeId
      in: path
      required: true
      schema: { type: string, pattern: "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$" }
...
    YearMonth:
      name: ym
      in: path
      required: true
      schema: { type: string, pattern: '^\d{4}-\d{2}$' }
      example: "2026-07"
...
  responses:
    Error:
      description: An error, at any non-2xx status.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: A short, stable machine-readable error code (not localized).
          examples: ["not_found"]
        message:
          type: string
          description: >-
            Optional diagnostic detail — which field failed validation and
            why, a decode error, and the like. Not stable, not localized,
            not meant to be pattern-matched by a client; `error` is the
            field to branch on.
...
    Date:
      type: string
      format: date
      pattern: '^\d{4}-\d{2}-\d{2}$'
      examples: ["2026-07-14"]
...
    Employee:
      type: object
      required: [id, tenantId, name, startDate]
      properties:
        id: { type: string, pattern: "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$" }
        tenantId: { type: string, pattern: "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$" }
        name: { type: string }
        startDate: { $ref: "#/components/schemas/Date" }
        endDate:
          oneOf:
            - $ref: "#/components/schemas/Date"
            - type: "null"
          description: "null means still employed."
    EmployeeCreate:
      type: object
      required: [name, startDate]
      properties:
        name: { type: string, minLength: 1 }
        startDate: { $ref: "#/components/schemas/Date" }
        endDate:
          oneOf:
            - $ref: "#/components/schemas/Date"
            - type: "null"
    EmployeeUpdate:
      type: object
      required: [name, startDate]
      properties:
        name: { type: string, minLength: 1 }
        startDate: { $ref: "#/components/schemas/Date" }
        endDate:
          oneOf:
            - $ref: "#/components/schemas/Date"
            - type: "null"
...
    SetHoursRequest:
      type: object
      required: [employeeId, date, hours]
      properties:
        employeeId: { type: string, pattern: "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$" }
        date: { $ref: "#/components/schemas/Date" }
        hours:
          type: string
          description: Decimal hours, `.` or `,` separator; "0" clears the cell.
    HoursResult:
      type: object
      required: [employeeTotal, dayTotal, grandTotal]
      properties:
        employeeTotal: { type: string }
        dayTotal: { type: string }
        grandTotal: { type: string }
...